Start the monitoring in the scheduler code instead of main
